+++ base_pack/flipper_share/flipper_share_app.h
@@ -9,6 +9,7 @@
 #include <gui/modules/dialog_ex.h>
 #include <gui/modules/submenu.h>
 #include <gui/modules/file_browser.h>
+#include <gui/modules/widget.h>
 #include <gui/view.h>
 #include <gui/view_dispatcher.h>
 #include <input/input.h>
@@ -38,7 +39,7 @@ struct FlipperShareApp {
     FileBrowser* file_browser;
     DialogEx* dialog_show_file;
     DialogEx* dialog_receive;
-    DialogEx* dialog_about;
+    Widget* widget_about;
 
     FuriString* result_path;
     char selected_file_path[256];